// src/api/generated/models/updateDevProfileRequest.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * brew
 * OpenAPI spec version: 0.0.1
 */
import type { SeniorityLevelEnum } from './seniorityLevelEnum';
import type { DevSpecialtyEnum } from './devSpecialtyEnum';

export interface UpdateDevProfileRequest {
  cpf?: string;
  /**
   * @minLength 3
   * @maxLength 255
   */
  name?: string;
  bio?: string;
  phone?: string;
  birthdate?: string;
  seniority_level?: SeniorityLevelEnum;
  specialty?: DevSpecialtyEnum;
}